+++ src/ttsim/__init__.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
-from ttsim.arg_templates import input_data, output
 from ttsim.interface_dag import main
 from ttsim.interface_dag_elements.shared import merge_trees
+from ttsim.main_args import input_data, output
 
 __all__ = [
     "input_data",
